Error Fixing still need to debug MORE

+++ Final Project C++/Final Project C++/Employee.h
@@ -6,24 +6,24 @@
 class Employee : public Contact
 {
 protected:
-	float	m_salaryPerHour;
+	int	m_salaryPerHour;
 	int		m_numOfWorkingHours;
 
 public:
-	Employee(const Contact& other, float salaryPerHour, int numOfWorkingHours = 0) : Contact(other), m_salaryPerHour(salaryPerHour), m_numOfWorkingHours(numOfWorkingHours) { }
+	Employee(const Contact& other, int salaryPerHour, int numOfWorkingHours = 0) : Contact(other), m_salaryPerHour(salaryPerHour), m_numOfWorkingHours(numOfWorkingHours) { }
 	Employee() { }
 
-	void setSalaryPerHour(float salaryPerHour)		 { m_salaryPerHour = salaryPerHour; }
+	void setSalaryPerHour(int salaryPerHour)		 { m_salaryPerHour = salaryPerHour; }
 	void setNumOfWorkingHours(int numOfWorkingHours) { m_numOfWorkingHours = numOfWorkingHours; }
 
-	float getSalaryPerHour()	const { return m_salaryPerHour; }
+	int getSalaryPerHour()	const { return m_salaryPerHour; }
 	int getNumOfWorkingHours()	const { return m_numOfWorkingHours; }
 
 	friend istream& operator>>(istream& in, Employee& employee);
 
 	virtual void toOs(ostream& os) const;
 
-	float calcSalary() const { return m_numOfWorkingHours*m_salaryPerHour; }
+	int calcSalary() const { return m_numOfWorkingHours*m_salaryPerHour; }
 };
 
 #endif
